@@ -35,12 +35,13 @@
     "cytoscape": "^3.2.16",
     "cytoscape-cose-bilkent": "^4.0.0",
     "cytoscape-edgehandles": "^3.5.1",
-    "esptk": "Nexus-Mods/node-esptk",
+    "esptk": "Nexus-Mods/node-esptk#electron7",
     "i18next": "^15.0.9",
     "immutability-helper": "^2.7.1",
     "js-yaml": "^3.12.0",
     "json-loader": "^0.5.7",
-    "loot": "Nexus-Mods/node-loot",
+    "loot": "Nexus-Mods/node-loot#electron7",
+    "markdown-ast": "^0.2.1",
     "prop-types": "^15.6.2",
     "react": "^16.8.6",
     "react-dnd": "^7.4.5",
@@ -49,8 +50,8 @@
     "react-select": "^1.2.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
-    "ts-loader": "4.4.2",
-    "vortex-api": "Nexus-Mods/vortex-api",
-    "webpack": "^4.16.5"
+    "ts-loader": "7.0.1",
+    "vortex-api": "Nexus-Mods/Vortex-api#electron7",
+    "webpack": "^4.42.1"
   }
 }
